Fix Bash and refactor
I tested this manually in ZSH and had pushed without making sure Bash
had no problems. Turns out it did, the command would complete and dump
the user back at the prompt before all the output was done.  Adding
cat at the end of the pipeline ensures that Bash will not return until
the subprocesses have completed. This seems to address the issue.

I've also centralized some of the repeated complexity within the
pipeline at the cost of simpler repetition.